msm8916-common: Separate fingerprints for variants
The build fingerprint is being constructed in a way that
will be valid for one of the unified device variants, but
will almost certainly be incorrect for all of the others.

Modify it so that each variant gets a build fingerprint
that corresponds exactly to an official ASUS release for
that variant.

Change-Id: I0f5e7ba03359f3c1f79cd576d28bd3d8ef1eef33